requirements.txt
tensorflow==2.8.2
pandas==1.5.0
Pillow==9.2.0
numpy==1.23.4